Restore events on genesis
Apparently, the concern was for genesis files which registered hundreds of
thousands of accounts, flooding the DB?

Requiring knowledge of the genesis to understand the blockchain state, as a
client, is extremely frustrating. Restoring events ensures any event-causing
behavior on genesis is easily accessible. While yes, this enables the DoS it
was removed to avoid, Serai doesn't have such a use case and therefore
shouldn't be at risk.