Remove a few pallets which definitely won't be used
PoW should be self-explanatory. Aura would also be removed if it wasn't used so
extensively in testing in this repo.

Multisig/recovery have off-chain cryptographic solutions which should be
promoted.

The atomic swap pallet is limited to Blake2, making it unusable with Bitcoin
and so on. While Substrate still supports adaptor signature swaps, those don't
require this pallet which likely only enables swaps with... other users of this
pallet? Which is too small a space to be justified.